mongodb/requirements.txt
fastapi
pymongo
python-jose
passlib
python-dotenv
pydantic
python-multipart